cidr: add option to return IP in CIDR notation
Certain providers require the IP address for a static config be in a
CIDR notation, as they do not have a separate `netmask` parameter.

Giving the user the option to choose if the returned IP address should
be in a CIDR notation or not reduces the complexity and "glue"
when mixing `cidrhost` and `split` to get the same result.

Signed-off-by: Tan Siewert <tan@siewert.io>